Add action types to popup location and delete previous action types
